Lesson 24

+++ app/Http/Controllers/ThreadsController.php
@@ -119,6 +119,9 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+
+        $this->authorize('update', $thread);
+        
         $thread->delete();
 
         if (request()->wantsJson()) {
